+++ client/src/App.jsx
@@ -5,14 +5,14 @@ import Login from "./views/Login/Login";
 import Clients from "./views/Clients/Clients";
 import Inventary from "./views/Inventary/Inventary";
 import Events from "./views/Events/Events";
-import EventDetails from "./views/EventDetails/EventDetails";
+import EventDetail from "./views/EventDetail/EventDetail";
 import AuthContextProvider from "./contexts/AuthContext";
 import PublicRoutes from "./components/router/PublicRoutes/PublicRoutes";
 import PrivateRoutes from "./components/router/PrivateRoutes/PrivateRoutes";
 import { Toaster } from "react-hot-toast";
 import Profile from "./views/Profile/Profile";
 import ClientDetail from "./views/ClientDetail/ClientDetail";
-import InventaryDetails from "./views/InventaryDetails/InventaryDetails";
+import InventaryDetail from "./views/InventaryDetail/InventaryDetail";
 import ModifyFurnitures from "./views/Furnitures/ModidyFurniture";
 
 function App() {
@@ -33,12 +33,12 @@ function App() {
               </Route>
               <Route path="inventario">
                 <Route index element={<Inventary />} />
-                <Route path=":id" element={<InventaryDetails />} />
+                <Route path=":id" element={<InventaryDetail />} />
               </Route>
               <Route path="profile" element={<Profile />} />
               <Route path="eventos">
                 <Route index element={<Events />} />
-                <Route path=":id" element={<EventDetails />} />
+                <Route path=":id" element={<EventDetail />} />
               </Route>
               <Route path="mobiliario" element={<ModifyFurnitures />} />
             </Route>
